Share logic between sprockets 2.x and 3.x processors
